oqs_sys::common
Function OQS_MEM_secure_free

pub unsafe extern "C" fn OQS_MEM_secure_free(
    ptr: *mut c_void,
    len: usize,
)

Zeros out len bytes of memory starting at ptr, then frees ptr.

Can be called with ptr = NULL, in which case no operation is performed.

Designed to be protected against optimizing compilers which try to remove “unnecessary” operations. Should be used for all buffers containing secret data.

@param[in] ptr The start of the memory to zero out and free. @param[in] len The number of bytes to zero out.