oqs_sys

Module kem

Re-exports

pub use super::common::OQS_STATUS;

Structs

OQS_KEM

Key encapsulation mechanism object

Constants

OQS_KEM_alg_bike_l1

OQS_KEM_alg_bike_l3

OQS_KEM_alg_bike_l5

OQS_KEM_alg_classic_mceliece_348864

OQS_KEM_alg_classic_mceliece_460896

OQS_KEM_alg_classic_mceliece_348864f

OQS_KEM_alg_classic_mceliece_460896f

OQS_KEM_alg_classic_mceliece_6688128

OQS_KEM_alg_classic_mceliece_6960119

OQS_KEM_alg_classic_mceliece_8192128

OQS_KEM_alg_classic_mceliece_6688128f

OQS_KEM_alg_classic_mceliece_6960119f

OQS_KEM_alg_classic_mceliece_8192128f

OQS_KEM_alg_efrodokem_640_aes

OQS_KEM_alg_efrodokem_640_shake

OQS_KEM_alg_efrodokem_976_aes

OQS_KEM_alg_efrodokem_976_shake

OQS_KEM_alg_efrodokem_1344_aes

OQS_KEM_alg_efrodokem_1344_shake

OQS_KEM_alg_frodokem_640_aes

OQS_KEM_alg_frodokem_640_shake

OQS_KEM_alg_frodokem_976_aes

OQS_KEM_alg_frodokem_976_shake

OQS_KEM_alg_frodokem_1344_aes

OQS_KEM_alg_frodokem_1344_shake

OQS_KEM_alg_hqc_1

OQS_KEM_alg_hqc_3

OQS_KEM_alg_hqc_5

OQS_KEM_alg_kyber_512

OQS_KEM_alg_kyber_768

OQS_KEM_alg_kyber_1024

OQS_KEM_alg_ml_kem_512

OQS_KEM_alg_ml_kem_768

OQS_KEM_alg_ml_kem_1024

OQS_KEM_alg_ntru_hps2048509

OQS_KEM_alg_ntru_hps2048677

OQS_KEM_alg_ntru_hps4096821

OQS_KEM_alg_ntru_hps40961229

OQS_KEM_alg_ntru_hrss701

OQS_KEM_alg_ntru_hrss1373

OQS_KEM_alg_ntruprime_sntrup761

OQS_KEM_algs_length

OQS_KEM_bike_l1_length_ciphertext

OQS_KEM_bike_l1_length_encaps_seed

OQS_KEM_bike_l1_length_keypair_seed

OQS_KEM_bike_l1_length_public_key

OQS_KEM_bike_l1_length_secret_key

OQS_KEM_bike_l1_length_shared_secret

OQS_KEM_bike_l3_length_ciphertext

OQS_KEM_bike_l3_length_encaps_seed

OQS_KEM_bike_l3_length_keypair_seed

OQS_KEM_bike_l3_length_public_key

OQS_KEM_bike_l3_length_secret_key

OQS_KEM_bike_l3_length_shared_secret

OQS_KEM_bike_l5_length_ciphertext

OQS_KEM_bike_l5_length_encaps_seed

OQS_KEM_bike_l5_length_keypair_seed

OQS_KEM_bike_l5_length_public_key

OQS_KEM_bike_l5_length_secret_key

OQS_KEM_bike_l5_length_shared_secret

OQS_KEM_classic_mceliece_348864_length_ciphertext

OQS_KEM_classic_mceliece_348864_length_encaps_seed

OQS_KEM_classic_mceliece_348864_length_keypair_seed

OQS_KEM_classic_mceliece_348864_length_public_key

OQS_KEM_classic_mceliece_348864_length_secret_key

OQS_KEM_classic_mceliece_348864_length_shared_secret

OQS_KEM_classic_mceliece_348864f_length_ciphertext

OQS_KEM_classic_mceliece_348864f_length_encaps_seed

OQS_KEM_classic_mceliece_348864f_length_keypair_seed

OQS_KEM_classic_mceliece_348864f_length_public_key

OQS_KEM_classic_mceliece_348864f_length_secret_key

OQS_KEM_classic_mceliece_348864f_length_shared_secret

OQS_KEM_classic_mceliece_460896_length_ciphertext

OQS_KEM_classic_mceliece_460896_length_encaps_seed

OQS_KEM_classic_mceliece_460896_length_keypair_seed

OQS_KEM_classic_mceliece_460896_length_public_key

OQS_KEM_classic_mceliece_460896_length_secret_key

OQS_KEM_classic_mceliece_460896_length_shared_secret

OQS_KEM_classic_mceliece_460896f_length_ciphertext

OQS_KEM_classic_mceliece_460896f_length_encaps_seed

OQS_KEM_classic_mceliece_460896f_length_keypair_seed

OQS_KEM_classic_mceliece_460896f_length_public_key

OQS_KEM_classic_mceliece_460896f_length_secret_key

OQS_KEM_classic_mceliece_460896f_length_shared_secret

OQS_KEM_classic_mceliece_6688128_length_ciphertext

OQS_KEM_classic_mceliece_6688128_length_encaps_seed

OQS_KEM_classic_mceliece_6688128_length_keypair_seed

OQS_KEM_classic_mceliece_6688128_length_public_key

OQS_KEM_classic_mceliece_6688128_length_secret_key

OQS_KEM_classic_mceliece_6688128_length_shared_secret

OQS_KEM_classic_mceliece_6688128f_length_ciphertext

OQS_KEM_classic_mceliece_6688128f_length_encaps_seed

OQS_KEM_classic_mceliece_6688128f_length_keypair_seed

OQS_KEM_classic_mceliece_6688128f_length_public_key

OQS_KEM_classic_mceliece_6688128f_length_secret_key

OQS_KEM_classic_mceliece_6688128f_length_shared_secret

OQS_KEM_classic_mceliece_6960119_length_ciphertext

OQS_KEM_classic_mceliece_6960119_length_encaps_seed

OQS_KEM_classic_mceliece_6960119_length_keypair_seed

OQS_KEM_classic_mceliece_6960119_length_public_key

OQS_KEM_classic_mceliece_6960119_length_secret_key

OQS_KEM_classic_mceliece_6960119_length_shared_secret

OQS_KEM_classic_mceliece_6960119f_length_ciphertext

OQS_KEM_classic_mceliece_6960119f_length_encaps_seed

OQS_KEM_classic_mceliece_6960119f_length_keypair_seed

OQS_KEM_classic_mceliece_6960119f_length_public_key

OQS_KEM_classic_mceliece_6960119f_length_secret_key

OQS_KEM_classic_mceliece_6960119f_length_shared_secret

OQS_KEM_classic_mceliece_8192128_length_ciphertext

OQS_KEM_classic_mceliece_8192128_length_encaps_seed

OQS_KEM_classic_mceliece_8192128_length_keypair_seed

OQS_KEM_classic_mceliece_8192128_length_public_key

OQS_KEM_classic_mceliece_8192128_length_secret_key

OQS_KEM_classic_mceliece_8192128_length_shared_secret

OQS_KEM_classic_mceliece_8192128f_length_ciphertext

OQS_KEM_classic_mceliece_8192128f_length_encaps_seed

OQS_KEM_classic_mceliece_8192128f_length_keypair_seed

OQS_KEM_classic_mceliece_8192128f_length_public_key

OQS_KEM_classic_mceliece_8192128f_length_secret_key

OQS_KEM_classic_mceliece_8192128f_length_shared_secret

OQS_KEM_efrodokem_640_aes_length_ciphertext

OQS_KEM_efrodokem_640_aes_length_encaps_seed

OQS_KEM_efrodokem_640_aes_length_keypair_seed

OQS_KEM_efrodokem_640_aes_length_public_key

OQS_KEM_efrodokem_640_aes_length_secret_key

OQS_KEM_efrodokem_640_aes_length_shared_secret

OQS_KEM_efrodokem_640_shake_length_ciphertext

OQS_KEM_efrodokem_640_shake_length_encaps_seed

OQS_KEM_efrodokem_640_shake_length_keypair_seed

OQS_KEM_efrodokem_640_shake_length_public_key

OQS_KEM_efrodokem_640_shake_length_secret_key

OQS_KEM_efrodokem_640_shake_length_shared_secret

OQS_KEM_efrodokem_976_aes_length_ciphertext

OQS_KEM_efrodokem_976_aes_length_encaps_seed

OQS_KEM_efrodokem_976_aes_length_keypair_seed

OQS_KEM_efrodokem_976_aes_length_public_key

OQS_KEM_efrodokem_976_aes_length_secret_key

OQS_KEM_efrodokem_976_aes_length_shared_secret

OQS_KEM_efrodokem_976_shake_length_ciphertext

OQS_KEM_efrodokem_976_shake_length_encaps_seed

OQS_KEM_efrodokem_976_shake_length_keypair_seed

OQS_KEM_efrodokem_976_shake_length_public_key

OQS_KEM_efrodokem_976_shake_length_secret_key

OQS_KEM_efrodokem_976_shake_length_shared_secret

OQS_KEM_efrodokem_1344_aes_length_ciphertext

OQS_KEM_efrodokem_1344_aes_length_encaps_seed

OQS_KEM_efrodokem_1344_aes_length_keypair_seed

OQS_KEM_efrodokem_1344_aes_length_public_key

OQS_KEM_efrodokem_1344_aes_length_secret_key

OQS_KEM_efrodokem_1344_aes_length_shared_secret

OQS_KEM_efrodokem_1344_shake_length_ciphertext

OQS_KEM_efrodokem_1344_shake_length_encaps_seed

OQS_KEM_efrodokem_1344_shake_length_keypair_seed

OQS_KEM_efrodokem_1344_shake_length_public_key

OQS_KEM_efrodokem_1344_shake_length_secret_key

OQS_KEM_efrodokem_1344_shake_length_shared_secret

OQS_KEM_frodokem_640_aes_length_ciphertext

OQS_KEM_frodokem_640_aes_length_encaps_seed

OQS_KEM_frodokem_640_aes_length_keypair_seed

OQS_KEM_frodokem_640_aes_length_public_key

OQS_KEM_frodokem_640_aes_length_secret_key

OQS_KEM_frodokem_640_aes_length_shared_secret

OQS_KEM_frodokem_640_shake_length_ciphertext

OQS_KEM_frodokem_640_shake_length_encaps_seed

OQS_KEM_frodokem_640_shake_length_keypair_seed

OQS_KEM_frodokem_640_shake_length_public_key

OQS_KEM_frodokem_640_shake_length_secret_key

OQS_KEM_frodokem_640_shake_length_shared_secret

OQS_KEM_frodokem_976_aes_length_ciphertext

OQS_KEM_frodokem_976_aes_length_encaps_seed

OQS_KEM_frodokem_976_aes_length_keypair_seed

OQS_KEM_frodokem_976_aes_length_public_key

OQS_KEM_frodokem_976_aes_length_secret_key

OQS_KEM_frodokem_976_aes_length_shared_secret

OQS_KEM_frodokem_976_shake_length_ciphertext

OQS_KEM_frodokem_976_shake_length_encaps_seed

OQS_KEM_frodokem_976_shake_length_keypair_seed

OQS_KEM_frodokem_976_shake_length_public_key

OQS_KEM_frodokem_976_shake_length_secret_key

OQS_KEM_frodokem_976_shake_length_shared_secret

OQS_KEM_frodokem_1344_aes_length_ciphertext

OQS_KEM_frodokem_1344_aes_length_encaps_seed

OQS_KEM_frodokem_1344_aes_length_keypair_seed

OQS_KEM_frodokem_1344_aes_length_public_key

OQS_KEM_frodokem_1344_aes_length_secret_key

OQS_KEM_frodokem_1344_aes_length_shared_secret

OQS_KEM_frodokem_1344_shake_length_ciphertext

OQS_KEM_frodokem_1344_shake_length_encaps_seed

OQS_KEM_frodokem_1344_shake_length_keypair_seed

OQS_KEM_frodokem_1344_shake_length_public_key

OQS_KEM_frodokem_1344_shake_length_secret_key

OQS_KEM_frodokem_1344_shake_length_shared_secret

OQS_KEM_hqc_1_length_ciphertext

OQS_KEM_hqc_1_length_encaps_seed

OQS_KEM_hqc_1_length_keypair_seed

OQS_KEM_hqc_1_length_public_key

OQS_KEM_hqc_1_length_secret_key

OQS_KEM_hqc_1_length_shared_secret

OQS_KEM_hqc_3_length_ciphertext

OQS_KEM_hqc_3_length_encaps_seed

OQS_KEM_hqc_3_length_keypair_seed

OQS_KEM_hqc_3_length_public_key

OQS_KEM_hqc_3_length_secret_key

OQS_KEM_hqc_3_length_shared_secret

OQS_KEM_hqc_5_length_ciphertext

OQS_KEM_hqc_5_length_encaps_seed

OQS_KEM_hqc_5_length_keypair_seed

OQS_KEM_hqc_5_length_public_key

OQS_KEM_hqc_5_length_secret_key

OQS_KEM_hqc_5_length_shared_secret

OQS_KEM_kyber_512_length_ciphertext

OQS_KEM_kyber_512_length_encaps_seed

OQS_KEM_kyber_512_length_keypair_seed

OQS_KEM_kyber_512_length_public_key

OQS_KEM_kyber_512_length_secret_key

OQS_KEM_kyber_512_length_shared_secret

OQS_KEM_kyber_768_length_ciphertext

OQS_KEM_kyber_768_length_encaps_seed

OQS_KEM_kyber_768_length_keypair_seed

OQS_KEM_kyber_768_length_public_key

OQS_KEM_kyber_768_length_secret_key

OQS_KEM_kyber_768_length_shared_secret

OQS_KEM_kyber_1024_length_ciphertext

OQS_KEM_kyber_1024_length_encaps_seed

OQS_KEM_kyber_1024_length_keypair_seed

OQS_KEM_kyber_1024_length_public_key

OQS_KEM_kyber_1024_length_secret_key

OQS_KEM_kyber_1024_length_shared_secret

OQS_KEM_ml_kem_512_length_ciphertext

OQS_KEM_ml_kem_512_length_encaps_seed

OQS_KEM_ml_kem_512_length_keypair_seed

OQS_KEM_ml_kem_512_length_public_key

OQS_KEM_ml_kem_512_length_secret_key

OQS_KEM_ml_kem_512_length_shared_secret

OQS_KEM_ml_kem_768_length_ciphertext

OQS_KEM_ml_kem_768_length_encaps_seed

OQS_KEM_ml_kem_768_length_keypair_seed

OQS_KEM_ml_kem_768_length_public_key

OQS_KEM_ml_kem_768_length_secret_key

OQS_KEM_ml_kem_768_length_shared_secret

OQS_KEM_ml_kem_1024_length_ciphertext

OQS_KEM_ml_kem_1024_length_encaps_seed

OQS_KEM_ml_kem_1024_length_keypair_seed

OQS_KEM_ml_kem_1024_length_public_key

OQS_KEM_ml_kem_1024_length_secret_key

OQS_KEM_ml_kem_1024_length_shared_secret

OQS_KEM_ntru_hps2048509_length_ciphertext

OQS_KEM_ntru_hps2048509_length_encaps_seed

OQS_KEM_ntru_hps2048509_length_keypair_seed

OQS_KEM_ntru_hps2048509_length_public_key

OQS_KEM_ntru_hps2048509_length_secret_key

OQS_KEM_ntru_hps2048509_length_shared_secret

OQS_KEM_ntru_hps2048677_length_ciphertext

OQS_KEM_ntru_hps2048677_length_encaps_seed

OQS_KEM_ntru_hps2048677_length_keypair_seed

OQS_KEM_ntru_hps2048677_length_public_key

OQS_KEM_ntru_hps2048677_length_secret_key

OQS_KEM_ntru_hps2048677_length_shared_secret

OQS_KEM_ntru_hps4096821_length_ciphertext

OQS_KEM_ntru_hps4096821_length_encaps_seed

OQS_KEM_ntru_hps4096821_length_keypair_seed

OQS_KEM_ntru_hps4096821_length_public_key

OQS_KEM_ntru_hps4096821_length_secret_key

OQS_KEM_ntru_hps4096821_length_shared_secret

OQS_KEM_ntru_hps40961229_length_ciphertext

OQS_KEM_ntru_hps40961229_length_encaps_seed

OQS_KEM_ntru_hps40961229_length_keypair_seed

OQS_KEM_ntru_hps40961229_length_public_key

OQS_KEM_ntru_hps40961229_length_secret_key

OQS_KEM_ntru_hps40961229_length_shared_secret

OQS_KEM_ntru_hrss701_length_ciphertext

OQS_KEM_ntru_hrss701_length_encaps_seed

OQS_KEM_ntru_hrss701_length_keypair_seed

OQS_KEM_ntru_hrss701_length_public_key

OQS_KEM_ntru_hrss701_length_secret_key

OQS_KEM_ntru_hrss701_length_shared_secret

OQS_KEM_ntru_hrss1373_length_ciphertext

OQS_KEM_ntru_hrss1373_length_encaps_seed

OQS_KEM_ntru_hrss1373_length_keypair_seed

OQS_KEM_ntru_hrss1373_length_public_key

OQS_KEM_ntru_hrss1373_length_secret_key

OQS_KEM_ntru_hrss1373_length_shared_secret

OQS_KEM_ntruprime_sntrup761_length_ciphertext

OQS_KEM_ntruprime_sntrup761_length_encaps_seed

OQS_KEM_ntruprime_sntrup761_length_keypair_seed

OQS_KEM_ntruprime_sntrup761_length_public_key

OQS_KEM_ntruprime_sntrup761_length_secret_key

OQS_KEM_ntruprime_sntrup761_length_shared_secret

Functions

OQS_KEM_alg_count⚠

Returns the number of key encapsulation mechanisms in liboqs. They can be enumerated with OQS_KEM_alg_identifier.

OQS_KEM_alg_identifier⚠

Returns identifiers for available key encapsulation mechanisms in liboqs. Used with OQS_KEM_new.

OQS_KEM_alg_is_enabled⚠

Indicates whether the specified algorithm was enabled at compile-time or not.

OQS_KEM_bike_l1_decaps⚠

OQS_KEM_bike_l1_encaps⚠

OQS_KEM_bike_l1_encaps_derand⚠

OQS_KEM_bike_l1_keypair⚠

OQS_KEM_bike_l1_keypair_derand⚠

OQS_KEM_bike_l1_new⚠

OQS_KEM_bike_l3_decaps⚠

OQS_KEM_bike_l3_encaps⚠

OQS_KEM_bike_l3_encaps_derand⚠

OQS_KEM_bike_l3_keypair⚠

OQS_KEM_bike_l3_keypair_derand⚠

OQS_KEM_bike_l3_new⚠

OQS_KEM_bike_l5_decaps⚠

OQS_KEM_bike_l5_encaps⚠

OQS_KEM_bike_l5_encaps_derand⚠

OQS_KEM_bike_l5_keypair⚠

OQS_KEM_bike_l5_keypair_derand⚠

OQS_KEM_bike_l5_new⚠

OQS_KEM_classic_mceliece_348864_decaps⚠

OQS_KEM_classic_mceliece_348864_encaps⚠

OQS_KEM_classic_mceliece_348864_encaps_derand⚠

OQS_KEM_classic_mceliece_348864_keypair⚠

OQS_KEM_classic_mceliece_348864_keypair_derand⚠

OQS_KEM_classic_mceliece_348864_new⚠

OQS_KEM_classic_mceliece_348864f_decaps⚠

OQS_KEM_classic_mceliece_348864f_encaps⚠

OQS_KEM_classic_mceliece_348864f_encaps_derand⚠

OQS_KEM_classic_mceliece_348864f_keypair⚠

OQS_KEM_classic_mceliece_348864f_keypair_derand⚠

OQS_KEM_classic_mceliece_348864f_new⚠

OQS_KEM_classic_mceliece_460896_decaps⚠

OQS_KEM_classic_mceliece_460896_encaps⚠

OQS_KEM_classic_mceliece_460896_encaps_derand⚠

OQS_KEM_classic_mceliece_460896_keypair⚠

OQS_KEM_classic_mceliece_460896_keypair_derand⚠

OQS_KEM_classic_mceliece_460896_new⚠

OQS_KEM_classic_mceliece_460896f_decaps⚠

OQS_KEM_classic_mceliece_460896f_encaps⚠

OQS_KEM_classic_mceliece_460896f_encaps_derand⚠

OQS_KEM_classic_mceliece_460896f_keypair⚠

OQS_KEM_classic_mceliece_460896f_keypair_derand⚠

OQS_KEM_classic_mceliece_460896f_new⚠

OQS_KEM_classic_mceliece_6688128_decaps⚠

OQS_KEM_classic_mceliece_6688128_encaps⚠

OQS_KEM_classic_mceliece_6688128_encaps_derand⚠

OQS_KEM_classic_mceliece_6688128_keypair⚠

OQS_KEM_classic_mceliece_6688128_keypair_derand⚠

OQS_KEM_classic_mceliece_6688128_new⚠

OQS_KEM_classic_mceliece_6688128f_decaps⚠

OQS_KEM_classic_mceliece_6688128f_encaps⚠

OQS_KEM_classic_mceliece_6688128f_encaps_derand⚠

OQS_KEM_classic_mceliece_6688128f_keypair⚠

OQS_KEM_classic_mceliece_6688128f_keypair_derand⚠

OQS_KEM_classic_mceliece_6688128f_new⚠

OQS_KEM_classic_mceliece_6960119_decaps⚠

OQS_KEM_classic_mceliece_6960119_encaps⚠

OQS_KEM_classic_mceliece_6960119_encaps_derand⚠

OQS_KEM_classic_mceliece_6960119_keypair⚠

OQS_KEM_classic_mceliece_6960119_keypair_derand⚠

OQS_KEM_classic_mceliece_6960119_new⚠

OQS_KEM_classic_mceliece_6960119f_decaps⚠

OQS_KEM_classic_mceliece_6960119f_encaps⚠

OQS_KEM_classic_mceliece_6960119f_encaps_derand⚠

OQS_KEM_classic_mceliece_6960119f_keypair⚠

OQS_KEM_classic_mceliece_6960119f_keypair_derand⚠

OQS_KEM_classic_mceliece_6960119f_new⚠

OQS_KEM_classic_mceliece_8192128_decaps⚠

OQS_KEM_classic_mceliece_8192128_encaps⚠

OQS_KEM_classic_mceliece_8192128_encaps_derand⚠

OQS_KEM_classic_mceliece_8192128_keypair⚠

OQS_KEM_classic_mceliece_8192128_keypair_derand⚠

OQS_KEM_classic_mceliece_8192128_new⚠

OQS_KEM_classic_mceliece_8192128f_decaps⚠

OQS_KEM_classic_mceliece_8192128f_encaps⚠

OQS_KEM_classic_mceliece_8192128f_encaps_derand⚠

OQS_KEM_classic_mceliece_8192128f_keypair⚠

OQS_KEM_classic_mceliece_8192128f_keypair_derand⚠

OQS_KEM_classic_mceliece_8192128f_new⚠

OQS_KEM_decaps⚠

Decapsulation algorithm.

OQS_KEM_efrodokem_640_aes_decaps⚠

OQS_KEM_efrodokem_640_aes_encaps⚠

OQS_KEM_efrodokem_640_aes_encaps_derand⚠

OQS_KEM_efrodokem_640_aes_keypair⚠

OQS_KEM_efrodokem_640_aes_keypair_derand⚠

OQS_KEM_efrodokem_640_aes_new⚠

OQS_KEM_efrodokem_640_shake_decaps⚠

OQS_KEM_efrodokem_640_shake_encaps⚠

OQS_KEM_efrodokem_640_shake_encaps_derand⚠

OQS_KEM_efrodokem_640_shake_keypair⚠

OQS_KEM_efrodokem_640_shake_keypair_derand⚠

OQS_KEM_efrodokem_640_shake_new⚠

OQS_KEM_efrodokem_976_aes_decaps⚠

OQS_KEM_efrodokem_976_aes_encaps⚠

OQS_KEM_efrodokem_976_aes_encaps_derand⚠

OQS_KEM_efrodokem_976_aes_keypair⚠

OQS_KEM_efrodokem_976_aes_keypair_derand⚠

OQS_KEM_efrodokem_976_aes_new⚠

OQS_KEM_efrodokem_976_shake_decaps⚠

OQS_KEM_efrodokem_976_shake_encaps⚠

OQS_KEM_efrodokem_976_shake_encaps_derand⚠

OQS_KEM_efrodokem_976_shake_keypair⚠

OQS_KEM_efrodokem_976_shake_keypair_derand⚠

OQS_KEM_efrodokem_976_shake_new⚠

OQS_KEM_efrodokem_1344_aes_decaps⚠

OQS_KEM_efrodokem_1344_aes_encaps⚠

OQS_KEM_efrodokem_1344_aes_encaps_derand⚠

OQS_KEM_efrodokem_1344_aes_keypair⚠

OQS_KEM_efrodokem_1344_aes_keypair_derand⚠

OQS_KEM_efrodokem_1344_aes_new⚠

OQS_KEM_efrodokem_1344_shake_decaps⚠

OQS_KEM_efrodokem_1344_shake_encaps⚠

OQS_KEM_efrodokem_1344_shake_encaps_derand⚠

OQS_KEM_efrodokem_1344_shake_keypair⚠

OQS_KEM_efrodokem_1344_shake_keypair_derand⚠

OQS_KEM_efrodokem_1344_shake_new⚠

OQS_KEM_encaps⚠

Encapsulation algorithm.

OQS_KEM_encaps_derand⚠

Derandomized encapsulation algorithm.

OQS_KEM_free⚠

Frees an OQS_KEM object that was constructed by OQS_KEM_new.

OQS_KEM_frodokem_640_aes_decaps⚠

OQS_KEM_frodokem_640_aes_encaps⚠

OQS_KEM_frodokem_640_aes_encaps_derand⚠

OQS_KEM_frodokem_640_aes_keypair⚠

OQS_KEM_frodokem_640_aes_keypair_derand⚠

OQS_KEM_frodokem_640_aes_new⚠

OQS_KEM_frodokem_640_shake_decaps⚠

OQS_KEM_frodokem_640_shake_encaps⚠

OQS_KEM_frodokem_640_shake_encaps_derand⚠

OQS_KEM_frodokem_640_shake_keypair⚠

OQS_KEM_frodokem_640_shake_keypair_derand⚠

OQS_KEM_frodokem_640_shake_new⚠

OQS_KEM_frodokem_976_aes_decaps⚠

OQS_KEM_frodokem_976_aes_encaps⚠

OQS_KEM_frodokem_976_aes_encaps_derand⚠

OQS_KEM_frodokem_976_aes_keypair⚠

OQS_KEM_frodokem_976_aes_keypair_derand⚠

OQS_KEM_frodokem_976_aes_new⚠

OQS_KEM_frodokem_976_shake_decaps⚠

OQS_KEM_frodokem_976_shake_encaps⚠

OQS_KEM_frodokem_976_shake_encaps_derand⚠

OQS_KEM_frodokem_976_shake_keypair⚠

OQS_KEM_frodokem_976_shake_keypair_derand⚠

OQS_KEM_frodokem_976_shake_new⚠

OQS_KEM_frodokem_1344_aes_decaps⚠

OQS_KEM_frodokem_1344_aes_encaps⚠

OQS_KEM_frodokem_1344_aes_encaps_derand⚠

OQS_KEM_frodokem_1344_aes_keypair⚠

OQS_KEM_frodokem_1344_aes_keypair_derand⚠

OQS_KEM_frodokem_1344_aes_new⚠

OQS_KEM_frodokem_1344_shake_decaps⚠

OQS_KEM_frodokem_1344_shake_encaps⚠

OQS_KEM_frodokem_1344_shake_encaps_derand⚠

OQS_KEM_frodokem_1344_shake_keypair⚠

OQS_KEM_frodokem_1344_shake_keypair_derand⚠

OQS_KEM_frodokem_1344_shake_new⚠

OQS_KEM_hqc_1_decaps⚠

OQS_KEM_hqc_1_encaps⚠

OQS_KEM_hqc_1_encaps_derand⚠

OQS_KEM_hqc_1_keypair⚠

OQS_KEM_hqc_1_keypair_derand⚠

OQS_KEM_hqc_1_new⚠

OQS_KEM_hqc_3_decaps⚠

OQS_KEM_hqc_3_encaps⚠

OQS_KEM_hqc_3_encaps_derand⚠

OQS_KEM_hqc_3_keypair⚠

OQS_KEM_hqc_3_keypair_derand⚠

OQS_KEM_hqc_3_new⚠

OQS_KEM_hqc_5_decaps⚠

OQS_KEM_hqc_5_encaps⚠

OQS_KEM_hqc_5_encaps_derand⚠

OQS_KEM_hqc_5_keypair⚠

OQS_KEM_hqc_5_keypair_derand⚠

OQS_KEM_hqc_5_new⚠

OQS_KEM_keypair⚠

Keypair generation algorithm.

OQS_KEM_keypair_derand⚠

Derandomized keypair generation algorithm.

OQS_KEM_kyber_512_decaps⚠

OQS_KEM_kyber_512_encaps⚠

OQS_KEM_kyber_512_encaps_derand⚠

OQS_KEM_kyber_512_keypair⚠

OQS_KEM_kyber_512_keypair_derand⚠

OQS_KEM_kyber_512_new⚠

OQS_KEM_kyber_768_decaps⚠

OQS_KEM_kyber_768_encaps⚠

OQS_KEM_kyber_768_encaps_derand⚠

OQS_KEM_kyber_768_keypair⚠

OQS_KEM_kyber_768_keypair_derand⚠

OQS_KEM_kyber_768_new⚠

OQS_KEM_kyber_1024_decaps⚠

OQS_KEM_kyber_1024_encaps⚠

OQS_KEM_kyber_1024_encaps_derand⚠

OQS_KEM_kyber_1024_keypair⚠

OQS_KEM_kyber_1024_keypair_derand⚠

OQS_KEM_kyber_1024_new⚠

OQS_KEM_ml_kem_512_decaps⚠

OQS_KEM_ml_kem_512_encaps⚠

OQS_KEM_ml_kem_512_encaps_derand⚠

OQS_KEM_ml_kem_512_keypair⚠

OQS_KEM_ml_kem_512_keypair_derand⚠

OQS_KEM_ml_kem_512_new⚠

OQS_KEM_ml_kem_768_decaps⚠

OQS_KEM_ml_kem_768_encaps⚠

OQS_KEM_ml_kem_768_encaps_derand⚠

OQS_KEM_ml_kem_768_keypair⚠

OQS_KEM_ml_kem_768_keypair_derand⚠

OQS_KEM_ml_kem_768_new⚠

OQS_KEM_ml_kem_1024_decaps⚠

OQS_KEM_ml_kem_1024_encaps⚠

OQS_KEM_ml_kem_1024_encaps_derand⚠

OQS_KEM_ml_kem_1024_keypair⚠

OQS_KEM_ml_kem_1024_keypair_derand⚠

OQS_KEM_ml_kem_1024_new⚠

OQS_KEM_new⚠

Constructs an OQS_KEM object for a particular algorithm.

OQS_KEM_ntru_hps2048509_decaps⚠

OQS_KEM_ntru_hps2048509_encaps⚠

OQS_KEM_ntru_hps2048509_encaps_derand⚠

OQS_KEM_ntru_hps2048509_keypair⚠

OQS_KEM_ntru_hps2048509_keypair_derand⚠

OQS_KEM_ntru_hps2048509_new⚠

OQS_KEM_ntru_hps2048677_decaps⚠

OQS_KEM_ntru_hps2048677_encaps⚠

OQS_KEM_ntru_hps2048677_encaps_derand⚠

OQS_KEM_ntru_hps2048677_keypair⚠

OQS_KEM_ntru_hps2048677_keypair_derand⚠

OQS_KEM_ntru_hps2048677_new⚠

OQS_KEM_ntru_hps4096821_decaps⚠

OQS_KEM_ntru_hps4096821_encaps⚠

OQS_KEM_ntru_hps4096821_encaps_derand⚠

OQS_KEM_ntru_hps4096821_keypair⚠

OQS_KEM_ntru_hps4096821_keypair_derand⚠

OQS_KEM_ntru_hps4096821_new⚠

OQS_KEM_ntru_hps40961229_decaps⚠

OQS_KEM_ntru_hps40961229_encaps⚠

OQS_KEM_ntru_hps40961229_encaps_derand⚠

OQS_KEM_ntru_hps40961229_keypair⚠

OQS_KEM_ntru_hps40961229_keypair_derand⚠

OQS_KEM_ntru_hps40961229_new⚠

OQS_KEM_ntru_hrss701_decaps⚠

OQS_KEM_ntru_hrss701_encaps⚠

OQS_KEM_ntru_hrss701_encaps_derand⚠

OQS_KEM_ntru_hrss701_keypair⚠

OQS_KEM_ntru_hrss701_keypair_derand⚠

OQS_KEM_ntru_hrss701_new⚠

OQS_KEM_ntru_hrss1373_decaps⚠

OQS_KEM_ntru_hrss1373_encaps⚠

OQS_KEM_ntru_hrss1373_encaps_derand⚠

OQS_KEM_ntru_hrss1373_keypair⚠

OQS_KEM_ntru_hrss1373_keypair_derand⚠

OQS_KEM_ntru_hrss1373_new⚠

OQS_KEM_ntruprime_sntrup761_decaps⚠

OQS_KEM_ntruprime_sntrup761_encaps⚠

OQS_KEM_ntruprime_sntrup761_encaps_derand⚠

OQS_KEM_ntruprime_sntrup761_keypair⚠

OQS_KEM_ntruprime_sntrup761_keypair_derand⚠

OQS_KEM_ntruprime_sntrup761_new⚠