Function oqs_sys::kem::OQS_KEM_keypair

pub unsafe extern "C" fn OQS_KEM_keypair(
    kem: *const OQS_KEM, 
    public_key: *mut u8, 
    secret_key: *mut u8
) -> OQS_STATUS

Keypair generation algorithm.

Caller is responsible for allocating sufficient memory for public_key and secret_key, based on the length_* members in this object or the per-scheme compile-time macros OQS_KEM_*_length_*.

@param[in] kem The OQS_KEM object representing the KEM. @param[out] public_key The public key represented as a byte string. @param[out] secret_key The secret key represented as a byte string. @return OQS_SUCCESS or OQS_ERROR