oqs_sys::kem
Function OQS_KEM_new

pub unsafe extern "C" fn OQS_KEM_new(
    method_name: *const c_char,
) -> *mut OQS_KEM

Constructs an OQS_KEM object for a particular algorithm.

Callers should always check whether the return value is NULL, which indicates either than an invalid algorithm name was provided, or that the requested algorithm was disabled at compile-time.

@param[in] method_name Name of the desired algorithm; one of the names in OQS_KEM_algs. @return An OQS_KEM for the particular algorithm, or NULL if the algorithm has been disabled at compile-time.