oqs_sys::rand

Function OQS_randombytes

pub unsafe extern "C" fn OQS_randombytes(
    random_array: *mut u8,
    bytes_to_read: usize,
)

Fills the given memory with the requested number of (pseudo)random bytes.

This implementation uses whichever algorithm has been selected by OQS_randombytes_switch_algorithm. The default is OQS_randombytes_system, which reads bytes from a system specific default source.

The caller is responsible for providing a buffer allocated with sufficient room.

@param[out] random_array Pointer to the memory to fill with (pseudo)random bytes @param[in] bytes_to_read The number of random bytes to read into memory