Function oqs_sys::sig::OQS_SIG_alg_identifier

pub unsafe extern "C" fn OQS_SIG_alg_identifier(i: usize) -> *const c_char

Returns identifiers for available signature schemes in liboqs. Used with OQS_SIG_new.

Note that algorithm identifiers are present in this list even when the algorithm is disabled at compile time.

@param[in] i Index of the algorithm identifier to return, 0 <= i < OQS_SIG_algs_length @return Algorithm identifier as a string, or NULL.