Function oqs_sys::sig::OQS_SIG_sign

pub unsafe extern "C" fn OQS_SIG_sign(
    sig: *const OQS_SIG, 
    signature: *mut u8, 
    signature_len: *mut usize, 
    message: *const u8, 
    message_len: usize, 
    secret_key: *const u8
) -> OQS_STATUS

Signature generation algorithm.

Caller is responsible for allocating sufficient memory for signnature, based on the length_* members in this object or the per-scheme compile-time macros OQS_SIG_*_length_*.

@param[in] sig The OQS_SIG object representing the signature scheme. @param[out] signature The signature on the message represented as a byte string. @param[out] signature_len The length of the signature. @param[in] message The message to sign represented as a byte string. @param[in] message_len The length of the message to sign. @param[in] secret_key The secret key represented as a byte string. @return OQS_SUCCESS or OQS_ERROR