Module oqs_sys::common

Enums

OQS_CPU_EXT
CPU runtime detection flags

OQS_STATUS
Represents return values from functions.

Constants

OQS_COMPILE_BUILD_TARGET

OQS_ENABLE_KEM_BIKE

OQS_ENABLE_KEM_CLASSIC_MCELIECE

OQS_ENABLE_KEM_FRODOKEM

OQS_ENABLE_KEM_HQC

OQS_ENABLE_KEM_KYBER

OQS_ENABLE_KEM_NTRU

OQS_ENABLE_KEM_NTRUPRIME

OQS_ENABLE_KEM_SABER

OQS_ENABLE_KEM_SIDH

OQS_ENABLE_KEM_SIKE

OQS_ENABLE_KEM_bike_l1

OQS_ENABLE_KEM_bike_l3

OQS_ENABLE_KEM_classic_mceliece_348864

OQS_ENABLE_KEM_classic_mceliece_348864_avx

OQS_ENABLE_KEM_classic_mceliece_348864f

OQS_ENABLE_KEM_classic_mceliece_348864f_avx

OQS_ENABLE_KEM_classic_mceliece_460896

OQS_ENABLE_KEM_classic_mceliece_460896_avx

OQS_ENABLE_KEM_classic_mceliece_460896f

OQS_ENABLE_KEM_classic_mceliece_460896f_avx

OQS_ENABLE_KEM_classic_mceliece_6688128

OQS_ENABLE_KEM_classic_mceliece_6688128_avx

OQS_ENABLE_KEM_classic_mceliece_6688128f

OQS_ENABLE_KEM_classic_mceliece_6688128f_avx

OQS_ENABLE_KEM_classic_mceliece_6960119

OQS_ENABLE_KEM_classic_mceliece_6960119_avx

OQS_ENABLE_KEM_classic_mceliece_6960119f

OQS_ENABLE_KEM_classic_mceliece_6960119f_avx

OQS_ENABLE_KEM_classic_mceliece_8192128

OQS_ENABLE_KEM_classic_mceliece_8192128_avx

OQS_ENABLE_KEM_classic_mceliece_8192128f

OQS_ENABLE_KEM_classic_mceliece_8192128f_avx

OQS_ENABLE_KEM_frodokem_640_aes

OQS_ENABLE_KEM_frodokem_640_shake

OQS_ENABLE_KEM_frodokem_976_aes

OQS_ENABLE_KEM_frodokem_976_shake

OQS_ENABLE_KEM_frodokem_1344_aes

OQS_ENABLE_KEM_frodokem_1344_shake

OQS_ENABLE_KEM_hqc_128

OQS_ENABLE_KEM_hqc_128_avx2

OQS_ENABLE_KEM_hqc_192

OQS_ENABLE_KEM_hqc_192_avx2

OQS_ENABLE_KEM_hqc_256

OQS_ENABLE_KEM_hqc_256_avx2

OQS_ENABLE_KEM_kyber_512

OQS_ENABLE_KEM_kyber_512_90s

OQS_ENABLE_KEM_kyber_512_90s_avx2

OQS_ENABLE_KEM_kyber_512_avx2

OQS_ENABLE_KEM_kyber_768

OQS_ENABLE_KEM_kyber_768_90s

OQS_ENABLE_KEM_kyber_768_90s_avx2

OQS_ENABLE_KEM_kyber_768_avx2

OQS_ENABLE_KEM_kyber_1024

OQS_ENABLE_KEM_kyber_1024_90s

OQS_ENABLE_KEM_kyber_1024_90s_avx2

OQS_ENABLE_KEM_kyber_1024_avx2

OQS_ENABLE_KEM_ntru_hps2048509

OQS_ENABLE_KEM_ntru_hps2048509_avx2

OQS_ENABLE_KEM_ntru_hps2048677

OQS_ENABLE_KEM_ntru_hps2048677_avx2

OQS_ENABLE_KEM_ntru_hps4096821

OQS_ENABLE_KEM_ntru_hps4096821_avx2

OQS_ENABLE_KEM_ntru_hrss701

OQS_ENABLE_KEM_ntru_hrss701_avx2

OQS_ENABLE_KEM_ntruprime_ntrulpr653

OQS_ENABLE_KEM_ntruprime_ntrulpr653_avx2

OQS_ENABLE_KEM_ntruprime_ntrulpr761

OQS_ENABLE_KEM_ntruprime_ntrulpr761_avx2

OQS_ENABLE_KEM_ntruprime_ntrulpr857

OQS_ENABLE_KEM_ntruprime_ntrulpr857_avx2

OQS_ENABLE_KEM_ntruprime_sntrup653

OQS_ENABLE_KEM_ntruprime_sntrup653_avx2

OQS_ENABLE_KEM_ntruprime_sntrup761

OQS_ENABLE_KEM_ntruprime_sntrup761_avx2

OQS_ENABLE_KEM_ntruprime_sntrup857

OQS_ENABLE_KEM_ntruprime_sntrup857_avx2

OQS_ENABLE_KEM_saber_firesaber

OQS_ENABLE_KEM_saber_firesaber_avx2

OQS_ENABLE_KEM_saber_lightsaber

OQS_ENABLE_KEM_saber_lightsaber_avx2

OQS_ENABLE_KEM_saber_saber

OQS_ENABLE_KEM_saber_saber_avx2

OQS_ENABLE_KEM_sidh_p434

OQS_ENABLE_KEM_sidh_p434_compressed

OQS_ENABLE_KEM_sidh_p503

OQS_ENABLE_KEM_sidh_p503_compressed

OQS_ENABLE_KEM_sidh_p610

OQS_ENABLE_KEM_sidh_p610_compressed

OQS_ENABLE_KEM_sidh_p751

OQS_ENABLE_KEM_sidh_p751_compressed

OQS_ENABLE_KEM_sike_p434

OQS_ENABLE_KEM_sike_p434_compressed

OQS_ENABLE_KEM_sike_p503

OQS_ENABLE_KEM_sike_p503_compressed

OQS_ENABLE_KEM_sike_p610

OQS_ENABLE_KEM_sike_p610_compressed

OQS_ENABLE_KEM_sike_p751

OQS_ENABLE_KEM_sike_p751_compressed

OQS_ENABLE_SHA3_xkcp_low_avx2

OQS_ENABLE_SIG_DILITHIUM

OQS_ENABLE_SIG_FALCON

OQS_ENABLE_SIG_PICNIC

OQS_ENABLE_SIG_RAINBOW

OQS_ENABLE_SIG_SPHINCS

OQS_ENABLE_SIG_dilithium_2

OQS_ENABLE_SIG_dilithium_2_aes

OQS_ENABLE_SIG_dilithium_2_aes_avx2

OQS_ENABLE_SIG_dilithium_2_avx2

OQS_ENABLE_SIG_dilithium_3

OQS_ENABLE_SIG_dilithium_3_aes

OQS_ENABLE_SIG_dilithium_3_aes_avx2

OQS_ENABLE_SIG_dilithium_3_avx2

OQS_ENABLE_SIG_dilithium_5

OQS_ENABLE_SIG_dilithium_5_aes

OQS_ENABLE_SIG_dilithium_5_aes_avx2

OQS_ENABLE_SIG_dilithium_5_avx2

OQS_ENABLE_SIG_falcon_512

OQS_ENABLE_SIG_falcon_512_avx2

OQS_ENABLE_SIG_falcon_1024

OQS_ENABLE_SIG_falcon_1024_avx2

OQS_ENABLE_SIG_picnic3_L1

OQS_ENABLE_SIG_picnic3_L3

OQS_ENABLE_SIG_picnic3_L5

OQS_ENABLE_SIG_picnic_L1_FS

OQS_ENABLE_SIG_picnic_L1_UR

OQS_ENABLE_SIG_picnic_L1_full

OQS_ENABLE_SIG_picnic_L3_FS

OQS_ENABLE_SIG_picnic_L3_UR

OQS_ENABLE_SIG_picnic_L3_full

OQS_ENABLE_SIG_picnic_L5_FS

OQS_ENABLE_SIG_picnic_L5_UR

OQS_ENABLE_SIG_picnic_L5_full

OQS_ENABLE_SIG_rainbow_III_circumzenithal

OQS_ENABLE_SIG_rainbow_III_classic

OQS_ENABLE_SIG_rainbow_III_compressed

OQS_ENABLE_SIG_rainbow_I_circumzenithal

OQS_ENABLE_SIG_rainbow_I_classic

OQS_ENABLE_SIG_rainbow_I_compressed

OQS_ENABLE_SIG_rainbow_V_circumzenithal

OQS_ENABLE_SIG_rainbow_V_classic

OQS_ENABLE_SIG_rainbow_V_compressed

OQS_ENABLE_SIG_sphincs_haraka_128f_robust

OQS_ENABLE_SIG_sphincs_haraka_128f_robust_aesni

OQS_ENABLE_SIG_sphincs_haraka_128f_simple

OQS_ENABLE_SIG_sphincs_haraka_128f_simple_aesni

OQS_ENABLE_SIG_sphincs_haraka_128s_robust

OQS_ENABLE_SIG_sphincs_haraka_128s_robust_aesni

OQS_ENABLE_SIG_sphincs_haraka_128s_simple

OQS_ENABLE_SIG_sphincs_haraka_128s_simple_aesni

OQS_ENABLE_SIG_sphincs_haraka_192f_robust

OQS_ENABLE_SIG_sphincs_haraka_192f_robust_aesni

OQS_ENABLE_SIG_sphincs_haraka_192f_simple

OQS_ENABLE_SIG_sphincs_haraka_192f_simple_aesni

OQS_ENABLE_SIG_sphincs_haraka_192s_robust

OQS_ENABLE_SIG_sphincs_haraka_192s_robust_aesni

OQS_ENABLE_SIG_sphincs_haraka_192s_simple

OQS_ENABLE_SIG_sphincs_haraka_192s_simple_aesni

OQS_ENABLE_SIG_sphincs_haraka_256f_robust

OQS_ENABLE_SIG_sphincs_haraka_256f_robust_aesni

OQS_ENABLE_SIG_sphincs_haraka_256f_simple

OQS_ENABLE_SIG_sphincs_haraka_256f_simple_aesni

OQS_ENABLE_SIG_sphincs_haraka_256s_robust

OQS_ENABLE_SIG_sphincs_haraka_256s_robust_aesni

OQS_ENABLE_SIG_sphincs_haraka_256s_simple

OQS_ENABLE_SIG_sphincs_haraka_256s_simple_aesni

OQS_ENABLE_SIG_sphincs_sha256_128f_robust

OQS_ENABLE_SIG_sphincs_sha256_128f_robust_avx2

OQS_ENABLE_SIG_sphincs_sha256_128f_simple

OQS_ENABLE_SIG_sphincs_sha256_128f_simple_avx2

OQS_ENABLE_SIG_sphincs_sha256_128s_robust

OQS_ENABLE_SIG_sphincs_sha256_128s_robust_avx2

OQS_ENABLE_SIG_sphincs_sha256_128s_simple

OQS_ENABLE_SIG_sphincs_sha256_128s_simple_avx2

OQS_ENABLE_SIG_sphincs_sha256_192f_robust

OQS_ENABLE_SIG_sphincs_sha256_192f_robust_avx2

OQS_ENABLE_SIG_sphincs_sha256_192f_simple

OQS_ENABLE_SIG_sphincs_sha256_192f_simple_avx2

OQS_ENABLE_SIG_sphincs_sha256_192s_robust

OQS_ENABLE_SIG_sphincs_sha256_192s_robust_avx2

OQS_ENABLE_SIG_sphincs_sha256_192s_simple

OQS_ENABLE_SIG_sphincs_sha256_192s_simple_avx2

OQS_ENABLE_SIG_sphincs_sha256_256f_robust

OQS_ENABLE_SIG_sphincs_sha256_256f_robust_avx2

OQS_ENABLE_SIG_sphincs_sha256_256f_simple

OQS_ENABLE_SIG_sphincs_sha256_256f_simple_avx2

OQS_ENABLE_SIG_sphincs_sha256_256s_robust

OQS_ENABLE_SIG_sphincs_sha256_256s_robust_avx2

OQS_ENABLE_SIG_sphincs_sha256_256s_simple

OQS_ENABLE_SIG_sphincs_sha256_256s_simple_avx2

OQS_ENABLE_SIG_sphincs_shake256_128f_robust

OQS_ENABLE_SIG_sphincs_shake256_128f_robust_avx2

OQS_ENABLE_SIG_sphincs_shake256_128f_simple

OQS_ENABLE_SIG_sphincs_shake256_128f_simple_avx2

OQS_ENABLE_SIG_sphincs_shake256_128s_robust

OQS_ENABLE_SIG_sphincs_shake256_128s_robust_avx2

OQS_ENABLE_SIG_sphincs_shake256_128s_simple

OQS_ENABLE_SIG_sphincs_shake256_128s_simple_avx2

OQS_ENABLE_SIG_sphincs_shake256_192f_robust

OQS_ENABLE_SIG_sphincs_shake256_192f_robust_avx2

OQS_ENABLE_SIG_sphincs_shake256_192f_simple

OQS_ENABLE_SIG_sphincs_shake256_192f_simple_avx2

OQS_ENABLE_SIG_sphincs_shake256_192s_robust

OQS_ENABLE_SIG_sphincs_shake256_192s_robust_avx2

OQS_ENABLE_SIG_sphincs_shake256_192s_simple

OQS_ENABLE_SIG_sphincs_shake256_192s_simple_avx2

OQS_ENABLE_SIG_sphincs_shake256_256f_robust

OQS_ENABLE_SIG_sphincs_shake256_256f_robust_avx2

OQS_ENABLE_SIG_sphincs_shake256_256f_simple

OQS_ENABLE_SIG_sphincs_shake256_256f_simple_avx2

OQS_ENABLE_SIG_sphincs_shake256_256s_robust

OQS_ENABLE_SIG_sphincs_shake256_256s_robust_avx2

OQS_ENABLE_SIG_sphincs_shake256_256s_simple

OQS_ENABLE_SIG_sphincs_shake256_256s_simple_avx2

OQS_USE_ADX_INSTRUCTIONS

OQS_USE_AES_INSTRUCTIONS

OQS_USE_AES_OPENSSL

OQS_USE_AVX2_INSTRUCTIONS

OQS_USE_AVX512_INSTRUCTIONS

OQS_USE_AVX_INSTRUCTIONS

OQS_USE_BMI1_INSTRUCTIONS

OQS_USE_BMI2_INSTRUCTIONS

OQS_USE_OPENSSL

OQS_USE_PCLMULQDQ_INSTRUCTIONS

OQS_USE_POPCNT_INSTRUCTIONS

OQS_USE_SHA2_OPENSSL

OQS_USE_SSE2_INSTRUCTIONS

OQS_USE_SSE3_INSTRUCTIONS

OQS_USE_SSE_INSTRUCTIONS

OQS_VERSION_TEXT

Functions

OQS_CPU_has_extension⚠
Checks if the CPU supports a given extension

OQS_MEM_aligned_alloc⚠
Internal implementation of C11 aligned_alloc to work around compiler quirks.

OQS_MEM_aligned_free⚠
Free memory allocated with OQS_MEM_aligned_alloc.

OQS_MEM_cleanse⚠
Zeros out len bytes of memory starting at ptr.

OQS_MEM_insecure_free⚠
Frees ptr.

OQS_MEM_secure_bcmp⚠
Constant time comparison of byte sequences a and b of length len. Returns 0 if the byte sequences are equal or if len=0. Returns 1 otherwise.

OQS_MEM_secure_free⚠
Zeros out len bytes of memory starting at ptr, then frees ptr.

OQS_init⚠
This currently only sets the values in the OQS_CPU_EXTENSIONS, and so has effect only when OQS_DIST_BUILD is set.