Module oqs_sys::common

Structs

OQS_CPU_EXTENSIONS

Enums

OQS_STATUS
Represents return values from functions.

Constants

OQS_COMPILE_BUILD_TARGET

OQS_ENABLE_KEM_FRODOKEM

OQS_ENABLE_KEM_frodokem_640_aes

OQS_ENABLE_SIG_DILITHIUM

OQS_ENABLE_SIG_dilithium_2

OQS_ENABLE_SIG_dilithium_2_avx2

OQS_USE_AES_INSTRUCTIONS

OQS_USE_AES_OPENSSL

OQS_USE_AVX2_INSTRUCTIONS

OQS_USE_AVX512_INSTRUCTIONS

OQS_USE_AVX_INSTRUCTIONS

OQS_USE_BMI1_INSTRUCTIONS

OQS_USE_BMI2_INSTRUCTIONS

OQS_USE_CPU_EXTENSIONS

OQS_USE_OPENSSL

OQS_USE_PCLMUL_INSTRUCTIONS

OQS_USE_POPCNT_INSTRUCTIONS

OQS_USE_SHA2_OPENSSL

OQS_USE_SHA3_OPENSSL

OQS_USE_SSE2_INSTRUCTIONS

OQS_USE_SSE3_INSTRUCTIONS

OQS_USE_SSE_INSTRUCTIONS

OQS_VERSION_TEXT

Functions

OQS_MEM_cleanse⚠
Zeros out len bytes of memory starting at ptr.

OQS_MEM_insecure_free⚠
Frees ptr.

OQS_MEM_secure_free⚠
Zeros out len bytes of memory starting at ptr, then frees ptr.

OQS_get_available_CPU_extensions⚠
Returns a list of available CPU extensions.

OQS_get_cpu_extension_name⚠
Returns name of available CPU extension.

OQS_init⚠
This currently only sets the values in the OQS_CPU_EXTENSIONS, and so has effect only when OQS_USE_CPU_EXTENSIONS and OQS_PORTABLE_BUILD are set.