Module oqs_sys::common

Enums

OQS_CPU_EXT

CPU runtime detection flags

OQS_STATUS

Represents return values from functions.

Constants

OQS_BUILD_ONLY_LIB

OQS_COMPILE_BUILD_TARGET

OQS_ENABLE_KEM_BIKE

OQS_ENABLE_KEM_CLASSIC_MCELIECE

OQS_ENABLE_KEM_FRODOKEM

OQS_ENABLE_KEM_HQC

OQS_ENABLE_KEM_KYBER

OQS_ENABLE_KEM_NTRUPRIME

OQS_ENABLE_KEM_bike_l1

OQS_ENABLE_KEM_bike_l3

OQS_ENABLE_KEM_bike_l5

OQS_ENABLE_KEM_classic_mceliece_348864

OQS_ENABLE_KEM_classic_mceliece_348864_avx2

OQS_ENABLE_KEM_classic_mceliece_348864f

OQS_ENABLE_KEM_classic_mceliece_348864f_avx2

OQS_ENABLE_KEM_classic_mceliece_460896

OQS_ENABLE_KEM_classic_mceliece_460896_avx2

OQS_ENABLE_KEM_classic_mceliece_460896f

OQS_ENABLE_KEM_classic_mceliece_460896f_avx2

OQS_ENABLE_KEM_classic_mceliece_6688128

OQS_ENABLE_KEM_classic_mceliece_6688128_avx2

OQS_ENABLE_KEM_classic_mceliece_6688128f

OQS_ENABLE_KEM_classic_mceliece_6688128f_avx2

OQS_ENABLE_KEM_classic_mceliece_6960119

OQS_ENABLE_KEM_classic_mceliece_6960119_avx2

OQS_ENABLE_KEM_classic_mceliece_6960119f

OQS_ENABLE_KEM_classic_mceliece_6960119f_avx2

OQS_ENABLE_KEM_classic_mceliece_8192128

OQS_ENABLE_KEM_classic_mceliece_8192128_avx2

OQS_ENABLE_KEM_classic_mceliece_8192128f

OQS_ENABLE_KEM_classic_mceliece_8192128f_avx2

OQS_ENABLE_KEM_frodokem_640_aes

OQS_ENABLE_KEM_frodokem_640_shake

OQS_ENABLE_KEM_frodokem_976_aes

OQS_ENABLE_KEM_frodokem_976_shake

OQS_ENABLE_KEM_frodokem_1344_aes

OQS_ENABLE_KEM_frodokem_1344_shake

OQS_ENABLE_KEM_hqc_128

OQS_ENABLE_KEM_hqc_128_avx2

OQS_ENABLE_KEM_hqc_192

OQS_ENABLE_KEM_hqc_192_avx2

OQS_ENABLE_KEM_hqc_256

OQS_ENABLE_KEM_hqc_256_avx2

OQS_ENABLE_KEM_kyber_512

OQS_ENABLE_KEM_kyber_512_avx2

OQS_ENABLE_KEM_kyber_768

OQS_ENABLE_KEM_kyber_768_avx2

OQS_ENABLE_KEM_kyber_1024

OQS_ENABLE_KEM_kyber_1024_avx2

OQS_ENABLE_KEM_ntruprime_sntrup761

OQS_ENABLE_KEM_ntruprime_sntrup761_avx2

OQS_ENABLE_SHA3_xkcp_low_avx2

OQS_ENABLE_SIG_DILITHIUM

OQS_ENABLE_SIG_FALCON

OQS_ENABLE_SIG_SPHINCS

OQS_ENABLE_SIG_dilithium_2

OQS_ENABLE_SIG_dilithium_2_avx2

OQS_ENABLE_SIG_dilithium_3

OQS_ENABLE_SIG_dilithium_3_avx2

OQS_ENABLE_SIG_dilithium_5

OQS_ENABLE_SIG_dilithium_5_avx2

OQS_ENABLE_SIG_falcon_512

OQS_ENABLE_SIG_falcon_512_avx2

OQS_ENABLE_SIG_falcon_1024

OQS_ENABLE_SIG_falcon_1024_avx2

OQS_ENABLE_SIG_sphincs_sha2_128f_simple

OQS_ENABLE_SIG_sphincs_sha2_128f_simple_avx2

OQS_ENABLE_SIG_sphincs_sha2_128s_simple

OQS_ENABLE_SIG_sphincs_sha2_128s_simple_avx2

OQS_ENABLE_SIG_sphincs_sha2_192f_simple

OQS_ENABLE_SIG_sphincs_sha2_192f_simple_avx2

OQS_ENABLE_SIG_sphincs_sha2_192s_simple

OQS_ENABLE_SIG_sphincs_sha2_192s_simple_avx2

OQS_ENABLE_SIG_sphincs_sha2_256f_simple

OQS_ENABLE_SIG_sphincs_sha2_256f_simple_avx2

OQS_ENABLE_SIG_sphincs_sha2_256s_simple

OQS_ENABLE_SIG_sphincs_sha2_256s_simple_avx2

OQS_ENABLE_SIG_sphincs_shake_128f_simple

OQS_ENABLE_SIG_sphincs_shake_128f_simple_avx2

OQS_ENABLE_SIG_sphincs_shake_128s_simple

OQS_ENABLE_SIG_sphincs_shake_128s_simple_avx2

OQS_ENABLE_SIG_sphincs_shake_192f_simple

OQS_ENABLE_SIG_sphincs_shake_192f_simple_avx2

OQS_ENABLE_SIG_sphincs_shake_192s_simple

OQS_ENABLE_SIG_sphincs_shake_192s_simple_avx2

OQS_ENABLE_SIG_sphincs_shake_256f_simple

OQS_ENABLE_SIG_sphincs_shake_256f_simple_avx2

OQS_ENABLE_SIG_sphincs_shake_256s_simple

OQS_ENABLE_SIG_sphincs_shake_256s_simple_avx2

OQS_OPT_TARGET

OQS_USE_ADX_INSTRUCTIONS

OQS_USE_AES_INSTRUCTIONS

OQS_USE_AVX2_INSTRUCTIONS

OQS_USE_AVX512_INSTRUCTIONS

OQS_USE_AVX_INSTRUCTIONS

OQS_USE_BMI1_INSTRUCTIONS

OQS_USE_BMI2_INSTRUCTIONS

OQS_USE_OPENSSL

OQS_USE_PCLMULQDQ_INSTRUCTIONS

OQS_USE_POPCNT_INSTRUCTIONS

OQS_USE_SHA2_OPENSSL

OQS_USE_SSE2_INSTRUCTIONS

OQS_USE_SSE3_INSTRUCTIONS

OQS_USE_SSE_INSTRUCTIONS

OQS_VERSION_TEXT

Functions

OQS_CPU_has_extension⚠

Checks if the CPU supports a given extension

OQS_MEM_aligned_alloc⚠

Internal implementation of C11 aligned_alloc to work around compiler quirks.

OQS_MEM_aligned_free⚠

Free memory allocated with OQS_MEM_aligned_alloc.

OQS_MEM_cleanse⚠

Zeros out len bytes of memory starting at ptr.

OQS_MEM_insecure_free⚠

Frees ptr.

OQS_MEM_secure_bcmp⚠

Constant time comparison of byte sequences a and b of length len. Returns 0 if the byte sequences are equal or if len=0. Returns 1 otherwise.

OQS_MEM_secure_free⚠

Zeros out len bytes of memory starting at ptr, then frees ptr.

OQS_destroy⚠

This function frees prefetched OpenSSL objects

OQS_init⚠

This currently sets the values in the OQS_CPU_EXTENSIONS and prefetches the OpenSSL objects if necessary.

OQS_version⚠

Return library version string.