Module oqs_sys::kem

Re-exports

pub use super::common::OQS_STATUS;

Structs

OQS_KEM
Key encapsulation mechanism object

Constants

OQS_KEM_alg_bike1_l1_cpa

OQS_KEM_alg_bike1_l1_fo

OQS_KEM_alg_bike1_l3_cpa

OQS_KEM_alg_bike1_l3_fo

OQS_KEM_alg_classic_mceliece_348864

OQS_KEM_alg_classic_mceliece_348864f

OQS_KEM_alg_classic_mceliece_460896

OQS_KEM_alg_classic_mceliece_460896f

OQS_KEM_alg_classic_mceliece_6688128

OQS_KEM_alg_classic_mceliece_6688128f

OQS_KEM_alg_classic_mceliece_6960119

OQS_KEM_alg_classic_mceliece_6960119f

OQS_KEM_alg_classic_mceliece_8192128

OQS_KEM_alg_classic_mceliece_8192128f

OQS_KEM_alg_default

OQS_KEM_alg_frodokem_640_aes

OQS_KEM_alg_frodokem_640_shake

OQS_KEM_alg_frodokem_976_aes

OQS_KEM_alg_frodokem_976_shake

OQS_KEM_alg_frodokem_1344_aes

OQS_KEM_alg_frodokem_1344_shake

OQS_KEM_alg_hqc_128

OQS_KEM_alg_hqc_192

OQS_KEM_alg_hqc_256

OQS_KEM_alg_kyber_512

OQS_KEM_alg_kyber_512_90s

OQS_KEM_alg_kyber_768

OQS_KEM_alg_kyber_768_90s

OQS_KEM_alg_kyber_1024

OQS_KEM_alg_kyber_1024_90s

OQS_KEM_alg_ntru_hps2048509

OQS_KEM_alg_ntru_hps2048677

OQS_KEM_alg_ntru_hps4096821

OQS_KEM_alg_ntru_hrss701

OQS_KEM_alg_ntruprime_ntrulpr653

OQS_KEM_alg_ntruprime_ntrulpr761

OQS_KEM_alg_ntruprime_ntrulpr857

OQS_KEM_alg_ntruprime_sntrup653

OQS_KEM_alg_ntruprime_sntrup761

OQS_KEM_alg_ntruprime_sntrup857

OQS_KEM_alg_saber_firesaber

OQS_KEM_alg_saber_lightsaber

OQS_KEM_alg_saber_saber

OQS_KEM_alg_sidh_p434

OQS_KEM_alg_sidh_p434_compressed

OQS_KEM_alg_sidh_p503

OQS_KEM_alg_sidh_p503_compressed

OQS_KEM_alg_sidh_p610

OQS_KEM_alg_sidh_p610_compressed

OQS_KEM_alg_sidh_p751

OQS_KEM_alg_sidh_p751_compressed

OQS_KEM_alg_sike_p434

OQS_KEM_alg_sike_p434_compressed

OQS_KEM_alg_sike_p503

OQS_KEM_alg_sike_p503_compressed

OQS_KEM_alg_sike_p610

OQS_KEM_alg_sike_p610_compressed

OQS_KEM_alg_sike_p751

OQS_KEM_alg_sike_p751_compressed

OQS_KEM_algs_length

OQS_KEM_frodokem_640_aes_length_ciphertext

OQS_KEM_frodokem_640_aes_length_public_key

OQS_KEM_frodokem_640_aes_length_secret_key

OQS_KEM_frodokem_640_aes_length_shared_secret

Functions

OQS_KEM_alg_count⚠
Returns the number of key encapsulation mechanisms in liboqs. They can be enumerated with OQS_KEM_alg_identifier.

OQS_KEM_alg_identifier⚠
Returns identifiers for available key encapsulation mechanisms in liboqs. Used with OQS_KEM_new.

OQS_KEM_alg_is_enabled⚠
Indicates whether the specified algorithm was enabled at compile-time or not.

OQS_KEM_decaps⚠
Decapsulation algorithm.

OQS_KEM_encaps⚠
Encapsulation algorithm.

OQS_KEM_free⚠
Frees an OQS_KEM object that was constructed by OQS_KEM_new.

OQS_KEM_frodokem_640_aes_decaps⚠

OQS_KEM_frodokem_640_aes_encaps⚠

OQS_KEM_frodokem_640_aes_keypair⚠

OQS_KEM_frodokem_640_aes_new⚠

OQS_KEM_keypair⚠
Keypair generation algorithm.

OQS_KEM_new⚠
Constructs an OQS_KEM object for a particular algorithm.