Module oqs_sys::kem

Re-exports

pub use super::common::OQS_STATUS;

Structs

OQS_KEM

Key encapsulation mechanism object

Constants

OQS_KEM_alg_bike_l1

OQS_KEM_alg_bike_l3

OQS_KEM_alg_classic_mceliece_348864

OQS_KEM_alg_classic_mceliece_348864f

OQS_KEM_alg_classic_mceliece_460896

OQS_KEM_alg_classic_mceliece_460896f

OQS_KEM_alg_classic_mceliece_6688128

OQS_KEM_alg_classic_mceliece_6688128f

OQS_KEM_alg_classic_mceliece_6960119

OQS_KEM_alg_classic_mceliece_6960119f

OQS_KEM_alg_classic_mceliece_8192128

OQS_KEM_alg_classic_mceliece_8192128f

OQS_KEM_alg_frodokem_640_aes

OQS_KEM_alg_frodokem_640_shake

OQS_KEM_alg_frodokem_976_aes

OQS_KEM_alg_frodokem_976_shake

OQS_KEM_alg_frodokem_1344_aes

OQS_KEM_alg_frodokem_1344_shake

OQS_KEM_alg_hqc_128

OQS_KEM_alg_hqc_192

OQS_KEM_alg_hqc_256

OQS_KEM_alg_kyber_512

OQS_KEM_alg_kyber_512_90s

OQS_KEM_alg_kyber_768

OQS_KEM_alg_kyber_768_90s

OQS_KEM_alg_kyber_1024

OQS_KEM_alg_kyber_1024_90s

OQS_KEM_alg_ntru_hps2048509

OQS_KEM_alg_ntru_hps2048677

OQS_KEM_alg_ntru_hps4096821

OQS_KEM_alg_ntru_hps40961229

OQS_KEM_alg_ntru_hrss701

OQS_KEM_alg_ntru_hrss1373

OQS_KEM_alg_ntruprime_ntrulpr653

OQS_KEM_alg_ntruprime_ntrulpr761

OQS_KEM_alg_ntruprime_ntrulpr857

OQS_KEM_alg_ntruprime_ntrulpr1277

OQS_KEM_alg_ntruprime_sntrup653

OQS_KEM_alg_ntruprime_sntrup761

OQS_KEM_alg_ntruprime_sntrup857

OQS_KEM_alg_ntruprime_sntrup1277

OQS_KEM_alg_saber_firesaber

OQS_KEM_alg_saber_lightsaber

OQS_KEM_alg_saber_saber

OQS_KEM_alg_sidh_p434

OQS_KEM_alg_sidh_p434_compressed

OQS_KEM_alg_sidh_p503

OQS_KEM_alg_sidh_p503_compressed

OQS_KEM_alg_sidh_p610

OQS_KEM_alg_sidh_p610_compressed

OQS_KEM_alg_sidh_p751

OQS_KEM_alg_sidh_p751_compressed

OQS_KEM_alg_sike_p434

OQS_KEM_alg_sike_p434_compressed

OQS_KEM_alg_sike_p503

OQS_KEM_alg_sike_p503_compressed

OQS_KEM_alg_sike_p610

OQS_KEM_alg_sike_p610_compressed

OQS_KEM_alg_sike_p751

OQS_KEM_alg_sike_p751_compressed

OQS_KEM_algs_length

OQS_KEM_bike_l1_length_ciphertext

OQS_KEM_bike_l1_length_public_key

OQS_KEM_bike_l1_length_secret_key

OQS_KEM_bike_l1_length_shared_secret

OQS_KEM_bike_l3_length_ciphertext

OQS_KEM_bike_l3_length_public_key

OQS_KEM_bike_l3_length_secret_key

OQS_KEM_bike_l3_length_shared_secret

OQS_KEM_classic_mceliece_348864_length_ciphertext

OQS_KEM_classic_mceliece_348864_length_public_key

OQS_KEM_classic_mceliece_348864_length_secret_key

OQS_KEM_classic_mceliece_348864_length_shared_secret

OQS_KEM_classic_mceliece_348864f_length_ciphertext

OQS_KEM_classic_mceliece_348864f_length_public_key

OQS_KEM_classic_mceliece_348864f_length_secret_key

OQS_KEM_classic_mceliece_348864f_length_shared_secret

OQS_KEM_classic_mceliece_460896_length_ciphertext

OQS_KEM_classic_mceliece_460896_length_public_key

OQS_KEM_classic_mceliece_460896_length_secret_key

OQS_KEM_classic_mceliece_460896_length_shared_secret

OQS_KEM_classic_mceliece_460896f_length_ciphertext

OQS_KEM_classic_mceliece_460896f_length_public_key

OQS_KEM_classic_mceliece_460896f_length_secret_key

OQS_KEM_classic_mceliece_460896f_length_shared_secret

OQS_KEM_classic_mceliece_6688128_length_ciphertext

OQS_KEM_classic_mceliece_6688128_length_public_key

OQS_KEM_classic_mceliece_6688128_length_secret_key

OQS_KEM_classic_mceliece_6688128_length_shared_secret

OQS_KEM_classic_mceliece_6688128f_length_ciphertext

OQS_KEM_classic_mceliece_6688128f_length_public_key

OQS_KEM_classic_mceliece_6688128f_length_secret_key

OQS_KEM_classic_mceliece_6688128f_length_shared_secret

OQS_KEM_classic_mceliece_6960119_length_ciphertext

OQS_KEM_classic_mceliece_6960119_length_public_key

OQS_KEM_classic_mceliece_6960119_length_secret_key

OQS_KEM_classic_mceliece_6960119_length_shared_secret

OQS_KEM_classic_mceliece_6960119f_length_ciphertext

OQS_KEM_classic_mceliece_6960119f_length_public_key

OQS_KEM_classic_mceliece_6960119f_length_secret_key

OQS_KEM_classic_mceliece_6960119f_length_shared_secret

OQS_KEM_classic_mceliece_8192128_length_ciphertext

OQS_KEM_classic_mceliece_8192128_length_public_key

OQS_KEM_classic_mceliece_8192128_length_secret_key

OQS_KEM_classic_mceliece_8192128_length_shared_secret

OQS_KEM_classic_mceliece_8192128f_length_ciphertext

OQS_KEM_classic_mceliece_8192128f_length_public_key

OQS_KEM_classic_mceliece_8192128f_length_secret_key

OQS_KEM_classic_mceliece_8192128f_length_shared_secret

OQS_KEM_frodokem_640_aes_length_ciphertext

OQS_KEM_frodokem_640_aes_length_public_key

OQS_KEM_frodokem_640_aes_length_secret_key

OQS_KEM_frodokem_640_aes_length_shared_secret

OQS_KEM_frodokem_640_shake_length_ciphertext

OQS_KEM_frodokem_640_shake_length_public_key

OQS_KEM_frodokem_640_shake_length_secret_key

OQS_KEM_frodokem_640_shake_length_shared_secret

OQS_KEM_frodokem_976_aes_length_ciphertext

OQS_KEM_frodokem_976_aes_length_public_key

OQS_KEM_frodokem_976_aes_length_secret_key

OQS_KEM_frodokem_976_aes_length_shared_secret

OQS_KEM_frodokem_976_shake_length_ciphertext

OQS_KEM_frodokem_976_shake_length_public_key

OQS_KEM_frodokem_976_shake_length_secret_key

OQS_KEM_frodokem_976_shake_length_shared_secret

OQS_KEM_frodokem_1344_aes_length_ciphertext

OQS_KEM_frodokem_1344_aes_length_public_key

OQS_KEM_frodokem_1344_aes_length_secret_key

OQS_KEM_frodokem_1344_aes_length_shared_secret

OQS_KEM_frodokem_1344_shake_length_ciphertext

OQS_KEM_frodokem_1344_shake_length_public_key

OQS_KEM_frodokem_1344_shake_length_secret_key

OQS_KEM_frodokem_1344_shake_length_shared_secret

OQS_KEM_hqc_128_length_ciphertext

OQS_KEM_hqc_128_length_public_key

OQS_KEM_hqc_128_length_secret_key

OQS_KEM_hqc_128_length_shared_secret

OQS_KEM_hqc_192_length_ciphertext

OQS_KEM_hqc_192_length_public_key

OQS_KEM_hqc_192_length_secret_key

OQS_KEM_hqc_192_length_shared_secret

OQS_KEM_hqc_256_length_ciphertext

OQS_KEM_hqc_256_length_public_key

OQS_KEM_hqc_256_length_secret_key

OQS_KEM_hqc_256_length_shared_secret

OQS_KEM_kyber_512_90s_length_ciphertext

OQS_KEM_kyber_512_90s_length_public_key

OQS_KEM_kyber_512_90s_length_secret_key

OQS_KEM_kyber_512_90s_length_shared_secret

OQS_KEM_kyber_512_length_ciphertext

OQS_KEM_kyber_512_length_public_key

OQS_KEM_kyber_512_length_secret_key

OQS_KEM_kyber_512_length_shared_secret

OQS_KEM_kyber_768_90s_length_ciphertext

OQS_KEM_kyber_768_90s_length_public_key

OQS_KEM_kyber_768_90s_length_secret_key

OQS_KEM_kyber_768_90s_length_shared_secret

OQS_KEM_kyber_768_length_ciphertext

OQS_KEM_kyber_768_length_public_key

OQS_KEM_kyber_768_length_secret_key

OQS_KEM_kyber_768_length_shared_secret

OQS_KEM_kyber_1024_90s_length_ciphertext

OQS_KEM_kyber_1024_90s_length_public_key

OQS_KEM_kyber_1024_90s_length_secret_key

OQS_KEM_kyber_1024_90s_length_shared_secret

OQS_KEM_kyber_1024_length_ciphertext

OQS_KEM_kyber_1024_length_public_key

OQS_KEM_kyber_1024_length_secret_key

OQS_KEM_kyber_1024_length_shared_secret

OQS_KEM_ntru_hps2048509_length_ciphertext

OQS_KEM_ntru_hps2048509_length_public_key

OQS_KEM_ntru_hps2048509_length_secret_key

OQS_KEM_ntru_hps2048509_length_shared_secret

OQS_KEM_ntru_hps2048677_length_ciphertext

OQS_KEM_ntru_hps2048677_length_public_key

OQS_KEM_ntru_hps2048677_length_secret_key

OQS_KEM_ntru_hps2048677_length_shared_secret

OQS_KEM_ntru_hps4096821_length_ciphertext

OQS_KEM_ntru_hps4096821_length_public_key

OQS_KEM_ntru_hps4096821_length_secret_key

OQS_KEM_ntru_hps4096821_length_shared_secret

OQS_KEM_ntru_hps40961229_length_ciphertext

OQS_KEM_ntru_hps40961229_length_public_key

OQS_KEM_ntru_hps40961229_length_secret_key

OQS_KEM_ntru_hps40961229_length_shared_secret

OQS_KEM_ntru_hrss701_length_ciphertext

OQS_KEM_ntru_hrss701_length_public_key

OQS_KEM_ntru_hrss701_length_secret_key

OQS_KEM_ntru_hrss701_length_shared_secret

OQS_KEM_ntru_hrss1373_length_ciphertext

OQS_KEM_ntru_hrss1373_length_public_key

OQS_KEM_ntru_hrss1373_length_secret_key

OQS_KEM_ntru_hrss1373_length_shared_secret

OQS_KEM_ntruprime_ntrulpr653_length_ciphertext

OQS_KEM_ntruprime_ntrulpr653_length_public_key

OQS_KEM_ntruprime_ntrulpr653_length_secret_key

OQS_KEM_ntruprime_ntrulpr653_length_shared_secret

OQS_KEM_ntruprime_ntrulpr761_length_ciphertext

OQS_KEM_ntruprime_ntrulpr761_length_public_key

OQS_KEM_ntruprime_ntrulpr761_length_secret_key

OQS_KEM_ntruprime_ntrulpr761_length_shared_secret

OQS_KEM_ntruprime_ntrulpr857_length_ciphertext

OQS_KEM_ntruprime_ntrulpr857_length_public_key

OQS_KEM_ntruprime_ntrulpr857_length_secret_key

OQS_KEM_ntruprime_ntrulpr857_length_shared_secret

OQS_KEM_ntruprime_ntrulpr1277_length_ciphertext

OQS_KEM_ntruprime_ntrulpr1277_length_public_key

OQS_KEM_ntruprime_ntrulpr1277_length_secret_key

OQS_KEM_ntruprime_ntrulpr1277_length_shared_secret

OQS_KEM_ntruprime_sntrup653_length_ciphertext

OQS_KEM_ntruprime_sntrup653_length_public_key

OQS_KEM_ntruprime_sntrup653_length_secret_key

OQS_KEM_ntruprime_sntrup653_length_shared_secret

OQS_KEM_ntruprime_sntrup761_length_ciphertext

OQS_KEM_ntruprime_sntrup761_length_public_key

OQS_KEM_ntruprime_sntrup761_length_secret_key

OQS_KEM_ntruprime_sntrup761_length_shared_secret

OQS_KEM_ntruprime_sntrup857_length_ciphertext

OQS_KEM_ntruprime_sntrup857_length_public_key

OQS_KEM_ntruprime_sntrup857_length_secret_key

OQS_KEM_ntruprime_sntrup857_length_shared_secret

OQS_KEM_ntruprime_sntrup1277_length_ciphertext

OQS_KEM_ntruprime_sntrup1277_length_public_key

OQS_KEM_ntruprime_sntrup1277_length_secret_key

OQS_KEM_ntruprime_sntrup1277_length_shared_secret

OQS_KEM_saber_firesaber_length_ciphertext

OQS_KEM_saber_firesaber_length_public_key

OQS_KEM_saber_firesaber_length_secret_key

OQS_KEM_saber_firesaber_length_shared_secret

OQS_KEM_saber_lightsaber_length_ciphertext

OQS_KEM_saber_lightsaber_length_public_key

OQS_KEM_saber_lightsaber_length_secret_key

OQS_KEM_saber_lightsaber_length_shared_secret

OQS_KEM_saber_saber_length_ciphertext

OQS_KEM_saber_saber_length_public_key

OQS_KEM_saber_saber_length_secret_key

OQS_KEM_saber_saber_length_shared_secret

OQS_KEM_sidh_p434_compressed_length_ciphertext

OQS_KEM_sidh_p434_compressed_length_public_key

OQS_KEM_sidh_p434_compressed_length_secret_key

OQS_KEM_sidh_p434_compressed_length_shared_secret

OQS_KEM_sidh_p434_length_ciphertext

OQS_KEM_sidh_p434_length_public_key

OQS_KEM_sidh_p434_length_secret_key

OQS_KEM_sidh_p434_length_shared_secret

OQS_KEM_sidh_p503_compressed_length_ciphertext

OQS_KEM_sidh_p503_compressed_length_public_key

OQS_KEM_sidh_p503_compressed_length_secret_key

OQS_KEM_sidh_p503_compressed_length_shared_secret

OQS_KEM_sidh_p503_length_ciphertext

OQS_KEM_sidh_p503_length_public_key

OQS_KEM_sidh_p503_length_secret_key

OQS_KEM_sidh_p503_length_shared_secret

OQS_KEM_sidh_p610_compressed_length_ciphertext

OQS_KEM_sidh_p610_compressed_length_public_key

OQS_KEM_sidh_p610_compressed_length_secret_key

OQS_KEM_sidh_p610_compressed_length_shared_secret

OQS_KEM_sidh_p610_length_ciphertext

OQS_KEM_sidh_p610_length_public_key

OQS_KEM_sidh_p610_length_secret_key

OQS_KEM_sidh_p610_length_shared_secret

OQS_KEM_sidh_p751_compressed_length_ciphertext

OQS_KEM_sidh_p751_compressed_length_public_key

OQS_KEM_sidh_p751_compressed_length_secret_key

OQS_KEM_sidh_p751_compressed_length_shared_secret

OQS_KEM_sidh_p751_length_ciphertext

OQS_KEM_sidh_p751_length_public_key

OQS_KEM_sidh_p751_length_secret_key

OQS_KEM_sidh_p751_length_shared_secret

OQS_KEM_sike_p434_compressed_length_ciphertext

OQS_KEM_sike_p434_compressed_length_public_key

OQS_KEM_sike_p434_compressed_length_secret_key

OQS_KEM_sike_p434_compressed_length_shared_secret

OQS_KEM_sike_p434_length_ciphertext

OQS_KEM_sike_p434_length_public_key

OQS_KEM_sike_p434_length_secret_key

OQS_KEM_sike_p434_length_shared_secret

OQS_KEM_sike_p503_compressed_length_ciphertext

OQS_KEM_sike_p503_compressed_length_public_key

OQS_KEM_sike_p503_compressed_length_secret_key

OQS_KEM_sike_p503_compressed_length_shared_secret

OQS_KEM_sike_p503_length_ciphertext

OQS_KEM_sike_p503_length_public_key

OQS_KEM_sike_p503_length_secret_key

OQS_KEM_sike_p503_length_shared_secret

OQS_KEM_sike_p610_compressed_length_ciphertext

OQS_KEM_sike_p610_compressed_length_public_key

OQS_KEM_sike_p610_compressed_length_secret_key

OQS_KEM_sike_p610_compressed_length_shared_secret

OQS_KEM_sike_p610_length_ciphertext

OQS_KEM_sike_p610_length_public_key

OQS_KEM_sike_p610_length_secret_key

OQS_KEM_sike_p610_length_shared_secret

OQS_KEM_sike_p751_compressed_length_ciphertext

OQS_KEM_sike_p751_compressed_length_public_key

OQS_KEM_sike_p751_compressed_length_secret_key

OQS_KEM_sike_p751_compressed_length_shared_secret

OQS_KEM_sike_p751_length_ciphertext

OQS_KEM_sike_p751_length_public_key

OQS_KEM_sike_p751_length_secret_key

OQS_KEM_sike_p751_length_shared_secret

Functions

OQS_KEM_alg_count⚠

Returns the number of key encapsulation mechanisms in liboqs. They can be enumerated with OQS_KEM_alg_identifier.

OQS_KEM_alg_identifier⚠

Returns identifiers for available key encapsulation mechanisms in liboqs. Used with OQS_KEM_new.

OQS_KEM_alg_is_enabled⚠

Indicates whether the specified algorithm was enabled at compile-time or not.

OQS_KEM_bike_l1_decaps⚠

OQS_KEM_bike_l1_encaps⚠

OQS_KEM_bike_l1_keypair⚠

OQS_KEM_bike_l1_new⚠

OQS_KEM_bike_l3_decaps⚠

OQS_KEM_bike_l3_encaps⚠

OQS_KEM_bike_l3_keypair⚠

OQS_KEM_bike_l3_new⚠

OQS_KEM_classic_mceliece_348864_decaps⚠

OQS_KEM_classic_mceliece_348864_encaps⚠

OQS_KEM_classic_mceliece_348864_keypair⚠

OQS_KEM_classic_mceliece_348864_new⚠

OQS_KEM_classic_mceliece_348864f_decaps⚠

OQS_KEM_classic_mceliece_348864f_encaps⚠

OQS_KEM_classic_mceliece_348864f_keypair⚠

OQS_KEM_classic_mceliece_348864f_new⚠

OQS_KEM_classic_mceliece_460896_decaps⚠

OQS_KEM_classic_mceliece_460896_encaps⚠

OQS_KEM_classic_mceliece_460896_keypair⚠

OQS_KEM_classic_mceliece_460896_new⚠

OQS_KEM_classic_mceliece_460896f_decaps⚠

OQS_KEM_classic_mceliece_460896f_encaps⚠

OQS_KEM_classic_mceliece_460896f_keypair⚠

OQS_KEM_classic_mceliece_460896f_new⚠

OQS_KEM_classic_mceliece_6688128_decaps⚠

OQS_KEM_classic_mceliece_6688128_encaps⚠

OQS_KEM_classic_mceliece_6688128_keypair⚠

OQS_KEM_classic_mceliece_6688128_new⚠

OQS_KEM_classic_mceliece_6688128f_decaps⚠

OQS_KEM_classic_mceliece_6688128f_encaps⚠

OQS_KEM_classic_mceliece_6688128f_keypair⚠

OQS_KEM_classic_mceliece_6688128f_new⚠

OQS_KEM_classic_mceliece_6960119_decaps⚠

OQS_KEM_classic_mceliece_6960119_encaps⚠

OQS_KEM_classic_mceliece_6960119_keypair⚠

OQS_KEM_classic_mceliece_6960119_new⚠

OQS_KEM_classic_mceliece_6960119f_decaps⚠

OQS_KEM_classic_mceliece_6960119f_encaps⚠

OQS_KEM_classic_mceliece_6960119f_keypair⚠

OQS_KEM_classic_mceliece_6960119f_new⚠

OQS_KEM_classic_mceliece_8192128_decaps⚠

OQS_KEM_classic_mceliece_8192128_encaps⚠

OQS_KEM_classic_mceliece_8192128_keypair⚠

OQS_KEM_classic_mceliece_8192128_new⚠

OQS_KEM_classic_mceliece_8192128f_decaps⚠

OQS_KEM_classic_mceliece_8192128f_encaps⚠

OQS_KEM_classic_mceliece_8192128f_keypair⚠

OQS_KEM_classic_mceliece_8192128f_new⚠

OQS_KEM_decaps⚠

Decapsulation algorithm.

OQS_KEM_encaps⚠

Encapsulation algorithm.

OQS_KEM_free⚠

Frees an OQS_KEM object that was constructed by OQS_KEM_new.

OQS_KEM_frodokem_640_aes_decaps⚠

OQS_KEM_frodokem_640_aes_encaps⚠

OQS_KEM_frodokem_640_aes_keypair⚠

OQS_KEM_frodokem_640_aes_new⚠

OQS_KEM_frodokem_640_shake_decaps⚠

OQS_KEM_frodokem_640_shake_encaps⚠

OQS_KEM_frodokem_640_shake_keypair⚠

OQS_KEM_frodokem_640_shake_new⚠

OQS_KEM_frodokem_976_aes_decaps⚠

OQS_KEM_frodokem_976_aes_encaps⚠

OQS_KEM_frodokem_976_aes_keypair⚠

OQS_KEM_frodokem_976_aes_new⚠

OQS_KEM_frodokem_976_shake_decaps⚠

OQS_KEM_frodokem_976_shake_encaps⚠

OQS_KEM_frodokem_976_shake_keypair⚠

OQS_KEM_frodokem_976_shake_new⚠

OQS_KEM_frodokem_1344_aes_decaps⚠

OQS_KEM_frodokem_1344_aes_encaps⚠

OQS_KEM_frodokem_1344_aes_keypair⚠

OQS_KEM_frodokem_1344_aes_new⚠

OQS_KEM_frodokem_1344_shake_decaps⚠

OQS_KEM_frodokem_1344_shake_encaps⚠

OQS_KEM_frodokem_1344_shake_keypair⚠

OQS_KEM_frodokem_1344_shake_new⚠

OQS_KEM_hqc_128_decaps⚠

OQS_KEM_hqc_128_encaps⚠

OQS_KEM_hqc_128_keypair⚠

OQS_KEM_hqc_128_new⚠

OQS_KEM_hqc_192_decaps⚠

OQS_KEM_hqc_192_encaps⚠

OQS_KEM_hqc_192_keypair⚠

OQS_KEM_hqc_192_new⚠

OQS_KEM_hqc_256_decaps⚠

OQS_KEM_hqc_256_encaps⚠

OQS_KEM_hqc_256_keypair⚠

OQS_KEM_hqc_256_new⚠

OQS_KEM_keypair⚠

Keypair generation algorithm.

OQS_KEM_kyber_512_90s_decaps⚠

OQS_KEM_kyber_512_90s_encaps⚠

OQS_KEM_kyber_512_90s_keypair⚠

OQS_KEM_kyber_512_90s_new⚠

OQS_KEM_kyber_512_decaps⚠

OQS_KEM_kyber_512_encaps⚠

OQS_KEM_kyber_512_keypair⚠

OQS_KEM_kyber_512_new⚠

OQS_KEM_kyber_768_90s_decaps⚠

OQS_KEM_kyber_768_90s_encaps⚠

OQS_KEM_kyber_768_90s_keypair⚠

OQS_KEM_kyber_768_90s_new⚠

OQS_KEM_kyber_768_decaps⚠

OQS_KEM_kyber_768_encaps⚠

OQS_KEM_kyber_768_keypair⚠

OQS_KEM_kyber_768_new⚠

OQS_KEM_kyber_1024_90s_decaps⚠

OQS_KEM_kyber_1024_90s_encaps⚠

OQS_KEM_kyber_1024_90s_keypair⚠

OQS_KEM_kyber_1024_90s_new⚠

OQS_KEM_kyber_1024_decaps⚠

OQS_KEM_kyber_1024_encaps⚠

OQS_KEM_kyber_1024_keypair⚠

OQS_KEM_kyber_1024_new⚠

OQS_KEM_new⚠

Constructs an OQS_KEM object for a particular algorithm.

OQS_KEM_ntru_hps2048509_decaps⚠

OQS_KEM_ntru_hps2048509_encaps⚠

OQS_KEM_ntru_hps2048509_keypair⚠

OQS_KEM_ntru_hps2048509_new⚠

OQS_KEM_ntru_hps2048677_decaps⚠

OQS_KEM_ntru_hps2048677_encaps⚠

OQS_KEM_ntru_hps2048677_keypair⚠

OQS_KEM_ntru_hps2048677_new⚠

OQS_KEM_ntru_hps4096821_decaps⚠

OQS_KEM_ntru_hps4096821_encaps⚠

OQS_KEM_ntru_hps4096821_keypair⚠

OQS_KEM_ntru_hps4096821_new⚠

OQS_KEM_ntru_hps40961229_decaps⚠

OQS_KEM_ntru_hps40961229_encaps⚠

OQS_KEM_ntru_hps40961229_keypair⚠

OQS_KEM_ntru_hps40961229_new⚠

OQS_KEM_ntru_hrss701_decaps⚠

OQS_KEM_ntru_hrss701_encaps⚠

OQS_KEM_ntru_hrss701_keypair⚠

OQS_KEM_ntru_hrss701_new⚠

OQS_KEM_ntru_hrss1373_decaps⚠

OQS_KEM_ntru_hrss1373_encaps⚠

OQS_KEM_ntru_hrss1373_keypair⚠

OQS_KEM_ntru_hrss1373_new⚠

OQS_KEM_ntruprime_ntrulpr653_decaps⚠

OQS_KEM_ntruprime_ntrulpr653_encaps⚠

OQS_KEM_ntruprime_ntrulpr653_keypair⚠

OQS_KEM_ntruprime_ntrulpr653_new⚠

OQS_KEM_ntruprime_ntrulpr761_decaps⚠

OQS_KEM_ntruprime_ntrulpr761_encaps⚠

OQS_KEM_ntruprime_ntrulpr761_keypair⚠

OQS_KEM_ntruprime_ntrulpr761_new⚠

OQS_KEM_ntruprime_ntrulpr857_decaps⚠

OQS_KEM_ntruprime_ntrulpr857_encaps⚠

OQS_KEM_ntruprime_ntrulpr857_keypair⚠

OQS_KEM_ntruprime_ntrulpr857_new⚠

OQS_KEM_ntruprime_ntrulpr1277_decaps⚠

OQS_KEM_ntruprime_ntrulpr1277_encaps⚠

OQS_KEM_ntruprime_ntrulpr1277_keypair⚠

OQS_KEM_ntruprime_ntrulpr1277_new⚠

OQS_KEM_ntruprime_sntrup653_decaps⚠

OQS_KEM_ntruprime_sntrup653_encaps⚠

OQS_KEM_ntruprime_sntrup653_keypair⚠

OQS_KEM_ntruprime_sntrup653_new⚠

OQS_KEM_ntruprime_sntrup761_decaps⚠

OQS_KEM_ntruprime_sntrup761_encaps⚠

OQS_KEM_ntruprime_sntrup761_keypair⚠

OQS_KEM_ntruprime_sntrup761_new⚠

OQS_KEM_ntruprime_sntrup857_decaps⚠

OQS_KEM_ntruprime_sntrup857_encaps⚠

OQS_KEM_ntruprime_sntrup857_keypair⚠

OQS_KEM_ntruprime_sntrup857_new⚠

OQS_KEM_ntruprime_sntrup1277_decaps⚠

OQS_KEM_ntruprime_sntrup1277_encaps⚠

OQS_KEM_ntruprime_sntrup1277_keypair⚠

OQS_KEM_ntruprime_sntrup1277_new⚠

OQS_KEM_saber_firesaber_decaps⚠

OQS_KEM_saber_firesaber_encaps⚠

OQS_KEM_saber_firesaber_keypair⚠

OQS_KEM_saber_firesaber_new⚠

OQS_KEM_saber_lightsaber_decaps⚠

OQS_KEM_saber_lightsaber_encaps⚠

OQS_KEM_saber_lightsaber_keypair⚠

OQS_KEM_saber_lightsaber_new⚠

OQS_KEM_saber_saber_decaps⚠

OQS_KEM_saber_saber_encaps⚠

OQS_KEM_saber_saber_keypair⚠

OQS_KEM_saber_saber_new⚠

OQS_KEM_sidh_p434_compressed_decaps⚠

OQS_KEM_sidh_p434_compressed_encaps⚠

OQS_KEM_sidh_p434_compressed_keypair⚠

OQS_KEM_sidh_p434_compressed_new⚠

OQS_KEM_sidh_p434_decaps⚠

OQS_KEM_sidh_p434_encaps⚠

OQS_KEM_sidh_p434_keypair⚠

OQS_KEM_sidh_p434_new⚠

OQS_KEM_sidh_p503_compressed_decaps⚠

OQS_KEM_sidh_p503_compressed_encaps⚠

OQS_KEM_sidh_p503_compressed_keypair⚠

OQS_KEM_sidh_p503_compressed_new⚠

OQS_KEM_sidh_p503_decaps⚠

OQS_KEM_sidh_p503_encaps⚠

OQS_KEM_sidh_p503_keypair⚠

OQS_KEM_sidh_p503_new⚠

OQS_KEM_sidh_p610_compressed_decaps⚠

OQS_KEM_sidh_p610_compressed_encaps⚠

OQS_KEM_sidh_p610_compressed_keypair⚠

OQS_KEM_sidh_p610_compressed_new⚠

OQS_KEM_sidh_p610_decaps⚠

OQS_KEM_sidh_p610_encaps⚠

OQS_KEM_sidh_p610_keypair⚠

OQS_KEM_sidh_p610_new⚠

OQS_KEM_sidh_p751_compressed_decaps⚠

OQS_KEM_sidh_p751_compressed_encaps⚠

OQS_KEM_sidh_p751_compressed_keypair⚠

OQS_KEM_sidh_p751_compressed_new⚠

OQS_KEM_sidh_p751_decaps⚠

OQS_KEM_sidh_p751_encaps⚠

OQS_KEM_sidh_p751_keypair⚠

OQS_KEM_sidh_p751_new⚠

OQS_KEM_sike_p434_compressed_decaps⚠

OQS_KEM_sike_p434_compressed_encaps⚠

OQS_KEM_sike_p434_compressed_keypair⚠

OQS_KEM_sike_p434_compressed_new⚠

OQS_KEM_sike_p434_decaps⚠

OQS_KEM_sike_p434_encaps⚠

OQS_KEM_sike_p434_keypair⚠

OQS_KEM_sike_p434_new⚠

OQS_KEM_sike_p503_compressed_decaps⚠

OQS_KEM_sike_p503_compressed_encaps⚠

OQS_KEM_sike_p503_compressed_keypair⚠

OQS_KEM_sike_p503_compressed_new⚠

OQS_KEM_sike_p503_decaps⚠

OQS_KEM_sike_p503_encaps⚠

OQS_KEM_sike_p503_keypair⚠

OQS_KEM_sike_p503_new⚠

OQS_KEM_sike_p610_compressed_decaps⚠

OQS_KEM_sike_p610_compressed_encaps⚠

OQS_KEM_sike_p610_compressed_keypair⚠

OQS_KEM_sike_p610_compressed_new⚠

OQS_KEM_sike_p610_decaps⚠

OQS_KEM_sike_p610_encaps⚠

OQS_KEM_sike_p610_keypair⚠

OQS_KEM_sike_p610_new⚠

OQS_KEM_sike_p751_compressed_decaps⚠

OQS_KEM_sike_p751_compressed_encaps⚠

OQS_KEM_sike_p751_compressed_keypair⚠

OQS_KEM_sike_p751_compressed_new⚠

OQS_KEM_sike_p751_decaps⚠

OQS_KEM_sike_p751_encaps⚠

OQS_KEM_sike_p751_keypair⚠

OQS_KEM_sike_p751_new⚠